# docs/openapi.yaml
openapi: "3.0.3"
...
info:
  title: Tiny Win API
  version: "1.0.0"
  description: |
    RESTful API for **Tiny Win** — the minimalist daily-win social platform.

    ### Core behaviours covered by this spec
    - **Auth** — register, login, logout, token refresh
    - **Wins** — create today's win (1 per user per day), retrieve own win, check status
    - **Feed** — friend feed (server-enforced lock until user has posted today)
    - **Reactions** — add / change / remove a reaction (🔥 👀 🤝)
    - **Streaks & Calendar** — user streak counters and win calendar

    ### Design principles
    - Server is source-of-truth for date boundaries (uses user's IANA timezone)
    - All write endpoints require `Idempotency-Key` header to prevent double-submit
    - Feed endpoint returns **403 FeedLocked** when user has not posted today
    - No public like/follower counts (anti-Instagram, PRD §2.3)

  contact:
    name: Tiny Win Engineering
    email: eng@tinywin.app
  license:
    name: Proprietary
servers:
  - url: https://api.tinywin.app/v1
    description: Production
  - url: https://staging-api.tinywin.app/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development

# ---------------------------------------------------------------------------
# Security Schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /auth/login` or `POST /auth/refresh`.
        Include in header: `Authorization: Bearer <token>`

  # -------------------------------------------------------------------------
  # Reusable Parameters
  # -------------------------------------------------------------------------
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID v4. Server stores this key and returns the cached
        response on duplicate requests within the same calendar day.
        Prevents double-submit (PRD §8, US-A3).
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
    WinId:
      name: win_id
      in: path
      required: true
      description: Unique identifier of the Tiny Win post.
      schema:
        type: string
        format: uuid
    UserId:
      name: user_id
      in: path
      required: true
      description: Unique identifier of the user.
      schema:
        type: string
        format: uuid

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, email, password, timezone]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          example: "hoanganh"
        email:
          type: string
          format: email
          example: "hoang@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "S3cur3P@ss!"
        timezone:
          type: string
          description: IANA timezone identifier
          example: "Asia/Ho_Chi_Minh"
        display_name:
          type: string
          maxLength: 50
          example: "Hoàng Anh"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "hoang@example.com"
        password:
          type: string
          format: password
          example: "S3cur3P@ss!"
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: Short-lived JWT (15 min)
          example: "eyJhbGciOiJSUzI1NiJ9..."
        refresh_token:
          type: string
          description: Long-lived opaque refresh token (30 days)
          example: "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token TTL in seconds
          example: 900
        user:
          $ref: '#/components/schemas/UserProfile'
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."

    # ── User ────────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        username:
          type: string
          example: "hoanganh"
        display_name:
          type: string
          example: "Hoàng Anh"
        timezone:
          type: string
          example: "Asia/Ho_Chi_Minh"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T07:00:00Z"

    # ── Wins ────────────────────────────────────────────────────────────────
...
    ReactionType:
      type: string
      enum: ["🔥", "👀", "🤝"]
      description: |
        Exactly three reaction types are supported (PRD §4.1, Rule #7):
        - 🔥 Fire
        - 👀 Eyes
        - 🤝 Handshake
    UpsertReactionRequest:
      type: object
      required: [type]
      properties:
        type:
          $ref: '#/components/schemas/ReactionType'
    ReactionResponse:
      type: object
      properties:
        win_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ReactionType'
          nullable: true
          description: Current reaction type, or null if removed.
        updated_at:
          type: string
          format: date-time

    # ── Feed ────────────────────────────────────────────────────────────────
...
    StreakResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        current_streak:
          type: integer
          description: Number of consecutive days with a win ending today or yesterday
          example: 7
        longest_streak:
          type: integer
          description: Longest ever consecutive-day streak
          example: 14
        total_wins:
          type: integer
          description: All-time total wins
          example: 42
        last_win_date:
          type: string
          format: date
          nullable: true
          example: "2024-07-15"
    CalendarResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        year:
          type: integer
          example: 2024
        month:
          type: integer
          minimum: 1
          maximum: 12
          example: 7
        days:
          type: array
          description: List of dates (in user timezone) when a win was posted this month.
          items:
            type: string
            format: date
          example: ["2024-07-01", "2024-07-02", "2024-07-05", "2024-07-15"]

    # ── Errors ──────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "WIN_ALREADY_EXISTS"
            message:
              type: string
              description: Human-readable explanation (calm, specific language per PRD §7)
              example: "Bạn đã đăng Tiny Win hôm nay rồi. Hẹn gặp lại ngày mai!"
            detail:
              type: string
              nullable: true
              description: Optional additional context
              example: "date_key=2024-07-15, existing_win_id=c3d4..."

  # -------------------------------------------------------------------------
  # Reusable Responses
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Token không hợp lệ hoặc đã hết hạn. Vui lòng đăng nhập lại."
...
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Không tìm thấy tài nguyên yêu cầu."
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Quá nhiều yêu cầu. Vui lòng thử lại sau."
    UnprocessableEntity:
      description: Validation error — request body failed business rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      description: |
        Creates a new Tiny Win account. Returns access & refresh tokens
        so the client is authenticated immediately after registration.
      security: []   # Public endpoint
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: "hoanganh"
              email: "hoang@example.com"
              password: "S3cur3P@ss!"
              timezone: "Asia/Ho_Chi_Minh"
              display_name: "Hoàng Anh"
      responses:
        "201":
          description: Account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "409":
          description: Email or username already taken.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "EMAIL_TAKEN"
                  message: "Email đã được đăng ký. Vui lòng dùng email khác hoặc đăng nhập."
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Returns JWT access token (15 min) and refresh token (30 days).
      security: []
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "INVALID_CREDENTIALS"
                  message: "Email hoặc mật khẩu không đúng."
        "429":
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange a valid refresh token for a new access token.
      security: []
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        "200":
          description: New access token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout — revoke refresh token
      description: |
        Invalidates the refresh token server-side (added to revocation list / Redis blocklist).
        Access token expires naturally (TTL 15 min).
      operationId: authLogout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        "204":
          description: Logged out successfully. No content returned.
        "401":
          $ref: '#/components/responses/Unauthorized'

  # ===========================================================================
  # WINS
  # ===========================================================================
...
  /wins/{win_id}/reaction:
    put:
      tags: [Reactions]
      summary: Add or change reaction on a win
      description: |
        Upserts the authenticated user's reaction on a specific win.

        **Rules (PRD §4.1, Business Rules #7-8):**
        - Only 3 reaction types: 🔥, 👀, 🤝
        - One reaction per user per win (UPSERT semantics — new type replaces old)
        - User cannot react to their own win (403)
        - Win author must be an accepted friend

        **Optimistic UI:** Client may update UI immediately and rollback on error.
        Server is source-of-truth; client must not show success before server confirms.
      operationId: upsertReaction
      parameters:
        - $ref: '#/components/parameters/WinId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertReactionRequest'
            example:
              type: "🔥"
      responses:
        "200":
          description: Reaction added or updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionResponse'
              example:
                win_id: "win-uuid-001"
                user_id: "user-uuid-me"
                type: "🔥"
                updated_at: "2024-07-15T09:00:00Z"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Cannot react to own win, or author is not an accepted friend.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "REACTION_FORBIDDEN"
                  message: "Bạn không thể reaction bài của chính mình."
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          description: Invalid reaction type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "INVALID_REACTION_TYPE"
                  message: "Chỉ chấp nhận ba loại reaction: 🔥, 👀, 🤝."

    delete:
      tags: [Reactions]
      summary: Remove reaction from a win
      description: |
        Removes the authenticated user's reaction from the specified win.
        Corresponds to "tap the same reaction again to toggle it off" (PRD §4.1, US-B3, Rule #9).
        Idempotent — deleting a non-existent reaction returns 204.
      operationId: deleteReaction
      parameters:
        - $ref: '#/components/parameters/WinId'
      responses:
        "204":
          description: Reaction removed (or did not exist). No content returned.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # STREAKS & CALENDAR
  # ===========================================================================
  /users/{user_id}/streaks:
    get:
      tags: [Streaks & Calendar]
      summary: Get streak data for a user
      description: |
        Returns current streak, longest streak and total win count.

        **Access rules:**
        - Any authenticated user can view their own streaks
        - Accepted friends can view each other's streaks
        - Streak counts are **not** publicly displayed in MVP UI (PRD §4.2)

        **Cache:** Cached in Redis per user (TTL 5 min). Invalidated on new win.
      operationId: getUserStreaks
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Streak data returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreakResponse'
              example:
                user_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                current_streak: 7
                longest_streak: 14
                total_wins: 42
                last_win_date: "2024-07-15"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Not authorized (not the user themselves or accepted friend).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          $ref: '#/components/responses/NotFound'
  /users/{user_id}/calendar:
    get:
      tags: [Streaks & Calendar]
      summary: Get win calendar for a user (month view)
      description: |
        Returns the list of dates in a given month when the user posted a Tiny Win.
        Used to render a calendar heatmap view.

        Dates are expressed in the **user's own timezone** (consistent with `date_key`).
        If `year` / `month` are omitted, returns the current calendar month.

        **Access rules:** Same as streaks — self or accepted friend only.
        **Cache:** Cached in Redis per (user, year, month) (TTL 10 min).
      operationId: getUserCalendar
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: year
          in: query
          required: false
          schema:
            type: integer
            minimum: 2024
            example: 2024
        - name: month
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 7
      responses:
        "200":
          description: Calendar data returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarResponse'
              example:
                user_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                year: 2024
                month: 7
                days: ["2024-07-01", "2024-07-02", "2024-07-05", "2024-07-15"]
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          description: Invalid year or month value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ===========================================================================
# Tags (for documentation grouping)
# ===========================================================================
...